+++ tlsh-cli/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+/** Provides the executable TLSH command-line application. */
+module io.github.camilyed.tlsh.cli {
+  requires info.picocli;
+  requires io.github.camilyed.tlsh;
+
+  opens io.github.camilyed.tlsh.cli to
+      info.picocli;
+}